[NVD-3343] change components place, now tab transfer parent for call transfer containter
