+++ src/ui/modules/work-section/modules/call/components/the-call.vue
@@ -44,7 +44,7 @@ import CallHeader from './call-header.vue';
 import Bridge from './call-merge/call-bridge-container.vue';
 import Numpad from './call-numpad/numpad.vue';
 import CallPreview from './call-preview.vue';
-import Transfer from './call-transfer/call-transfer-container.vue';
+import Transfer from './call-transfer/tabs/call-transfer-tabs.vue';
 
 export default {
   name: 'TheCall',
